chore(tests): consolidate dotenv integration tests into env_test.go
