+++ integration/env_test.go
@@ -86,3 +86,143 @@ func TestEnv_SucceedsInGlobalModeWithoutTrust(t *testing.T) {
 		t.Errorf("expected trust check to be bypassed in global mode, got: %s", stderr)
 	}
 }
+
+// TestDotEnv_InjectsValueIntoWindsorEnv verifies that a plain KEY=VALUE line in
+// contexts/<ctx>/.env is exported by `windsor env`. Bare `windsor init` (used by
+// PrepareFixture) provisions the "local" context.
+func TestDotEnv_InjectsValueIntoWindsorEnv(t *testing.T) {
+	t.Parallel()
+	dir, env := helpers.PrepareFixture(t, "default")
+
+	dotEnvPath := filepath.Join(dir, "contexts", "local", ".env")
+	if err := os.WriteFile(dotEnvPath, []byte("HYPERV_HOST=hyperv.local\n"), 0600); err != nil {
+		t.Fatalf("write .env fixture: %v", err)
+	}
+
+	stdout, stderr, err := helpers.RunCLI(dir, []string{"env"}, env)
+	if err != nil {
+		t.Fatalf("env: %v\nstderr: %s", err, stderr)
+	}
+	if !strings.Contains(string(stdout), "HYPERV_HOST=hyperv.local") {
+		t.Errorf("expected HYPERV_HOST=hyperv.local in output, got:\n%s", stdout)
+	}
+}
+
+// TestDotEnv_AbsentFileDoesNotBreakWindsorEnv verifies that `windsor env` succeeds
+// normally when no contexts/<ctx>/.env file is present.
+func TestDotEnv_AbsentFileDoesNotBreakWindsorEnv(t *testing.T) {
+	t.Parallel()
+	dir, env := helpers.PrepareFixture(t, "default")
+
+	_, stderr, err := helpers.RunCLI(dir, []string{"env"}, env)
+	if err != nil {
+		t.Fatalf("expected env to succeed with no .env file, got error: %v\nstderr: %s", err, stderr)
+	}
+}
+
+// TestDotEnv_ContextEnvironmentKeyOverridesDotEnv verifies that the declarative
+// contexts.<ctx>.environment key in windsor.yaml takes precedence over the same
+// key supplied via contexts/<ctx>/.env, matching the documented precedence where
+// .env is the lowest-precedence base layer.
+func TestDotEnv_ContextEnvironmentKeyOverridesDotEnv(t *testing.T) {
+	t.Parallel()
+	dir, env := helpers.CopyFixtureOnly(t, "default")
+
+	windsorYamlPath := filepath.Join(dir, "windsor.yaml")
+	windsorYaml := `version: v1alpha1
+contexts:
+  local:
+    environment:
+      SHARED_VAR: from-windsor-yaml
+`
+	if err := os.WriteFile(windsorYamlPath, []byte(windsorYaml), 0644); err != nil {
+		t.Fatalf("write windsor.yaml: %v", err)
+	}
+	helpers.MarkAsGitRepo(t, dir)
+	if _, stderr, err := helpers.RunCLI(dir, []string{"init"}, env); err != nil {
+		t.Fatalf("windsor init: %v\nstderr: %s", err, stderr)
+	}
+
+	dotEnvPath := filepath.Join(dir, "contexts", "local", ".env")
+	if err := os.WriteFile(dotEnvPath, []byte("SHARED_VAR=from-dotenv\n"), 0600); err != nil {
+		t.Fatalf("write .env fixture: %v", err)
+	}
+
+	stdout, stderr, err := helpers.RunCLI(dir, []string{"env"}, env)
+	if err != nil {
+		t.Fatalf("env: %v\nstderr: %s", err, stderr)
+	}
+	if !strings.Contains(string(stdout), "SHARED_VAR=from-windsor-yaml") {
+		t.Errorf("expected the windsor.yaml environment key to win over .env, got:\n%s", stdout)
+	}
+	if strings.Contains(string(stdout), "from-dotenv") {
+		t.Errorf("expected .env value to be overridden, got:\n%s", stdout)
+	}
+}
+
+// TestDotEnv_ContextGitignoreCoversDotEnvFile verifies that `windsor init` writes
+// contexts/<ctx>/.gitignore covering .env so the file can never be committed.
+func TestDotEnv_ContextGitignoreCoversDotEnvFile(t *testing.T) {
+	t.Parallel()
+	dir, _ := helpers.PrepareFixture(t, "default")
+
+	gitignorePath := filepath.Join(dir, "contexts", "local", ".gitignore")
+	content, err := os.ReadFile(gitignorePath)
+	if err != nil {
+		t.Fatalf("expected contexts/local/.gitignore to exist: %v", err)
+	}
+	if !strings.Contains(string(content), ".env") {
+		t.Errorf("expected .gitignore to cover .env, got:\n%s", content)
+	}
+}
+
+// TestTerraformDotEnv_InjectedWhenCdIntoTerraformDirectory verifies that a value in
+// contexts/<ctx>/terraform/.env reaches `windsor env` when the operator is cd'd into
+// a Terraform component directory — the same TerraformProvider.GetEnvVars call that
+// `windsor up`/`apply`/`plan`/`destroy` use per component.
+func TestTerraformDotEnv_InjectedWhenCdIntoTerraformDirectory(t *testing.T) {
+	t.Parallel()
+	dir, env := helpers.PrepareFixture(t, "terraform-dotenv")
+
+	dotEnvPath := filepath.Join(dir, "contexts", "local", "terraform", ".env")
+	if err := os.MkdirAll(filepath.Dir(dotEnvPath), 0750); err != nil {
+		t.Fatalf("mkdir terraform dir: %v", err)
+	}
+	if err := os.WriteFile(dotEnvPath, []byte("HYPERV_HOST=hyperv.local\n"), 0600); err != nil {
+		t.Fatalf("write terraform/.env fixture: %v", err)
+	}
+
+	componentDir := filepath.Join(dir, "terraform", "null")
+	stdout, stderr, err := helpers.RunCLI(componentDir, []string{"env"}, env)
+	if err != nil {
+		t.Fatalf("env: %v\nstderr: %s", err, stderr)
+	}
+	if !strings.Contains(string(stdout), "HYPERV_HOST=hyperv.local") {
+		t.Errorf("expected HYPERV_HOST=hyperv.local in output, got:\n%s", stdout)
+	}
+}
+
+// TestTerraformDotEnv_AbsentFromHookOutsideTerraformDirectory verifies that a value in
+// contexts/<ctx>/terraform/.env does NOT reach `windsor env` when the operator is not
+// cd'd into a Terraform directory — the core performance property: content that's only
+// relevant to Terraform work never touches the general interactive hook path.
+func TestTerraformDotEnv_AbsentFromHookOutsideTerraformDirectory(t *testing.T) {
+	t.Parallel()
+	dir, env := helpers.PrepareFixture(t, "terraform-dotenv")
+
+	dotEnvPath := filepath.Join(dir, "contexts", "local", "terraform", ".env")
+	if err := os.MkdirAll(filepath.Dir(dotEnvPath), 0750); err != nil {
+		t.Fatalf("mkdir terraform dir: %v", err)
+	}
+	if err := os.WriteFile(dotEnvPath, []byte("HYPERV_HOST=hyperv.local\n"), 0600); err != nil {
+		t.Fatalf("write terraform/.env fixture: %v", err)
+	}
+
+	stdout, stderr, err := helpers.RunCLI(dir, []string{"env"}, env)
+	if err != nil {
+		t.Fatalf("env: %v\nstderr: %s", err, stderr)
+	}
+	if strings.Contains(string(stdout), "hyperv.local") {
+		t.Errorf("expected HYPERV_HOST to be absent outside a Terraform directory, got:\n%s", stdout)
+	}
+}
